session_retry method moved into miopen/utils
