+++ tuna/miopen/worker/fin_class.py
@@ -49,7 +49,7 @@ from tuna.miopen.worker.fin_utils import compose_config_obj
 from tuna.miopen.worker.fin_utils import get_fin_slv_status
 from tuna.miopen.utils.config_type import ConfigType
 from tuna.miopen.utils.parsing import parse_pdb_key
-from tuna.utils.db_utility import session_retry
+from tuna.miopen.utils.miopen_db_utils import session_retry
 from tuna.miopen.utils.miopen_db_utils import get_solver_ids, get_id_solvers
 from tuna.utils.db_utility import gen_select_objs, gen_insert_query, gen_update_query
 from tuna.utils.db_utility import get_class_by_tablename, has_attr_set
