+++ tuna/rocmlir/rocmlir_worker.py
@@ -39,7 +39,8 @@ from sqlalchemy.inspection import inspect
 from tuna.dbBase.sql_alchemy import DbSession
 from tuna.worker_interface import WorkerInterface
 from tuna.rocmlir.rocmlir_tables import RocMLIRDBTables
-from tuna.utils.db_utility import session_retry, gen_insert_query
+from tuna.utils.db_utility import gen_insert_query
+from tuna.miopen.utils.miopen_db_utils import session_retry
 from tuna.rocmlir.config_type import ConfigType
