+++ tuna/worker_interface.py
@@ -51,7 +51,7 @@ from tuna.machine import Machine
 from tuna.abort import chk_abort_file
 from tuna.utils.metadata import TUNA_LOG_DIR, NUM_SQL_RETRIES, MAX_JOB_RETRIES, LOG_TIMEOUT
 from tuna.tables_interface import DBTablesInterface
-from tuna.utils.db_utility import session_retry
+from tuna.miopen.utils.miopen_db_utils import session_retry
 from tuna.utils.db_utility import gen_select_objs, gen_update_query, has_attr_set, connect_db
 from tuna.connection import Connection
 from tuna.utils.utility import SimpleDict
